moved files needed for the form gen outside of app dir | removed unused dev helpers
