+++ components/stepper.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react"
 import Stepper from "./stepper"
-import { allSections } from "../app/create/creator-sections"
+import { allSections } from "../form-generator/generator-sections"
 
 jest.mock("next/navigation", () => ({
   useParams: () => ({
